Cast the type of the base address of the processor-specific data structure from integer to u64 .

+++ src/kernel/arch/x86_64/arch.c
@@ -364,7 +364,7 @@ this_cpu(void)
 {
     struct p_data *pdata;
 
-    pdata = (struct p_data *)(P_DATA_BASE + lapic_id() * P_DATA_SIZE);
+    pdata = (struct p_data *)((u64)P_DATA_BASE + lapic_id() * P_DATA_SIZE);
 
     return pdata;
 }
